2024/10/12 11:28
Muitas coisas funcionando

+++ Assets/Scripts/HUD/PlayerLifeManager.cs
@@ -10,6 +10,6 @@ public class PlayerLifeManager : MonoBehaviour
 
     public void UpdateLife(int playercurrentHP, int playerMaxHP)
     {
-        playerLifeBar.size =  playercurrentHP / playerMaxHP;
+        playerLifeBar.size =  (float)playercurrentHP / playerMaxHP;
     }
 }
